fix(codegen): do not generate sourcemap for empty length spans (#11892)
Upstream may have modified and generated an incorrect span.
